Update whitespace handling of `grid` shortcode
Strips leading whitespace of the content so markdown blocks
get processed properly.

Adds the necessary space at the start and end of the content,
as well as before and after the generated HTML, which allows
the heavy lifting to be done by the regular rendering from 11ty.
This also avoids messing with whitespace deeper in the content,
though we're unlikely to have many.