Use Nodemon to watch deep imports of 11ty's config
Eleventy's configured to watch files in `eleventy` and reload the configuration.
It reloads the configuration both when files imported directly by the `eleventy.config.js`
change, and when the `import`s of these files change.

However, when reloading, code changes of the deeper imports are not reflected.
Eleventy keeps running the code that was loaded for the first run of a deeper import.
This is documented in this issue of Eleventy: https://github.com/11ty/eleventy/issues/3899

To circumvent that, we can restart Eleventy fully by using Nodemon to watch the files
loaded by Eleventy's configuration, thankfully all gathered in our `eleventy` folder.
When they change, Nodemon can then restart the whole Eleventy server, in a whole new process,
rather than having it keep running in the same process and just reloading its configuration.

In terms of overhead, Eleventy's reset of the configuration was already triggering a rebuild
of all the files, like restarting it does. There is a little overhead from shutting down
and re-starting a process, but it's barely noticeable.

Content changes (inside `src`) are left to be handled by Eleventy, which could allow us to use incremental build in the future.