Add `insideHTML` option to `blockShortcode`
When inside an HTML block, the blank line injected around shortcode's output
were problematic as it made Markdown it interpret the output as markdown.

In the case of the `video` shortcode, this lead to the `<video>` tags
being wrapped in `<p>`, adding extra margin at their bottom.

This PR adds the same `insideHTML` option as the `pairedBlockShortcode`
to `blockShortcode` allowing to control whether or not blank lines
should be wrapping the output.