Allow pages to be excluded from the Service Navigation in regular pages
In anticipation of having an Accessibility statement, allows pages to have an `excludeFromNavigation` property. When that property is set to `false`, the service navigation will not include the page.

The implementation repurposes the `pages` collection into a `serviceNavigationPages` collection, as the only purpose of `collections.pages` was to create the Service Navigation.