Display content on the left when there's no side navigation
Introduce a new `app-page-wrapper--no-navigation` setting a one column
layout when no side navigation is present (like on the Overview page,
or the upcoming Accessibility Statement page).

Width of the wrapper is capped to the width the content is displayed
at when a side navigation is present. This provides consistency on desktop
while allowing the content to take as much space as necessary before that.