Add sitemap specific titles
Adds an optional `sitemapTitle` property to pages' frontmatter.
It allows to set a different title for use on the sitemap pages,
so that pages with the same title in different sections of the site
can be distinguished by users (when displayed in Voice Over rotor's list of links
for example).

Sets the property on the relevant pages.